:twisted_rightwards_arrows: 깃 팀 추방 머지, 테스트 통과

+++ server/src/controllers/github/kickTeam.js
@@ -1,41 +1,40 @@
 require("dotenv").config();
-const FAIL = "github_invite_fail";
-const SUCCESS = "github_invite_success";
+
+const FAIL = "github_kick_fail";
+const SUCCESS = "github_kick_success";
+
 const axios = require("axios");
+const token = process.env.GITHUB_TOKEN;
 module.exports = {
-  //1. req = cohort, userArray(변경 예정)
-  //2. 클라이언트에서 req를 보낸다.
-  //3. req에서 보내 준 데이터와 일치하는 (아마 fidnAll?) 유저 정보 DB에서 가져온다.
-  //4. kickUserFromTeam 함수 실행해서 쫓아낸다 cohort: full-immersive-24th
-  //5. res로 userId, userName, userGithubName, status를 보낸다.(<< 이건 왜...?)
-  post: (userdata) => {
-    const token = process.env.GITHUB_TOKEN;
-    async function kickUserFromTeam() {
-      try {
-        let fromTmp = userdata.log
-          .split(",")[1]
-          .toLowerCase()
-          .split(" ")
-          .join("-");
-        const from = fromTmp.replace("기", "th");
-        const username = userdata.githubUserName;
-        await axios({
-          method: "delete",
-          headers: {
-            Authorization: `token ${token}`,
-          },
-          url: `https://api.github.com/orgs/codestates/${from}/test/memberships/${username}`,
-        });
-        return new Promise((resolve, reject) => {
-          resolve(SUCCESS);
-        });
-      } catch (err) {
-        console.log(err);
-        return new Promise((resolve, reject) => {
-          resolve(FAIL);
-        });
-      }
+  post: async (userdata) => {
+    
+    if (userdata.message) {
+      return new Promise((resolve, reject) => {
+        resolve(FAIL);
+      });
     }
-    kickUserFromTeam();
-  },
+    try {
+      //full immersive full pre 인경우 핸들링 필요
+      // let fromTmp = userdata.log.split(",")[1].toLowerCase().split(" ").join("-");
+      // const from = fromTmp.replace("기", "th");
+      const from = userdata.cohort;		//테스트용 
+      const username = userdata.githubUserName;
+
+      await axios({
+        method: "delete",
+        headers: {
+          Authorization: `token ${token}`,
+        },
+        url: `https://api.github.com/orgs/codestates/teams/${from}/memberships/${username}`,
+      });
+      return new Promise((resolve, reject) => {
+        resolve(SUCCESS);
+      });
+    } catch (err) {
+      console.log(err.message);
+      return new Promise((resolve, reject) => {
+        resolve(FAIL);
+      });
+    }
+  }
 };
